pyproject.toml
[project]
name = "langchain-udemy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
	"python-dotenv>=1.0.0",
	"langchain>=1.2.15",
	"langchain-ollama>=0.2.0",
	"langchain-openai==1.1.14",
	"langchain-tavily>=0.2.18",
	"tavily-python>=0.7.23",
	"openai>=2.26.0,<3",
	"scikit-learn>=1.5.0",
	"joblib>=1.3.0",
	"fastapi>=0.115.0",
	"uvicorn[standard]>=0.30.0",
	"streamlit>=1.36.0",
]
